requirements.txt
asgiref
defusedxml
diff-match-patch
Django
django-adminactions
django-crispy-forms
django-import-export
django-jazzmin
django-widget-tweaks
et-xmlfile
MarkupPy
odfpy
openpyxl
pytz
PyYAML
sqlparse
tablib
xlrd
xlwt
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
